RUMM-610 Add benchmark test for Data Uploader's memory usage
